+++ semana10/LabeX/labex/src/components/HeaderDashBord/HeaderDashBord.jsx
@@ -1,7 +1,7 @@
-import React, { useEffect } from 'react'
-import styled from 'styled-components'
+import React from 'react'
 import { useHistory } from 'react-router-dom';
 import { goToDashBord, goToLogin, goToDashBordForm} from '../../Router/goToPages';
+import {Header, NavMenu, Link, BntLogout} from "./styled"
 
 const HeaderDashBord = (props) => {
      const history = useHistory()
@@ -19,54 +19,3 @@ const HeaderDashBord = (props) => {
 }
 
 export default HeaderDashBord
-
-const Header = styled.header`
-background-color: #66f;
-height: 100px;
-width: 100%;
-padding: 10px;
-color: #fff;
-display: flex;
-justify-content: space-between;
-align-items: flex-end;
-position: relative;
-`
-
-const BntLogout = styled.button`
-    padding: 10px;
-    background-color: #868686;
-    color: #fff;
-    font-size: 16px;
-    border: none;
-    cursor: pointer;
-    position: absolute;
-    top: 10px;
-    right: 10px;
-    border-radius: 10px;
-    :hover {
-        color: #F00;
-        background-color: #fff;
-    }
-`
-
-
-const NavMenu = styled.div`
-display: flex;
-justify-content: space-around;
-align-items: flex-end;
-width: 30%;
-height: 50%;
-
-`
-
-const Link = styled.p`
-font-size:22px;
-padding: 3px;
-color: #fff;
-text-decoration: none;
-cursor: pointer;
-:hover {
-    background: #fff;
-    color: #000;
-}
-`
