+++ semana10/LabeX/labex/src/components/TripCard.jsx/TripCard.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import styled from 'styled-components'
+import {Card,Info} from './styled'
 
 
 const TripCard = (props) => {
@@ -18,36 +18,3 @@ const TripCard = (props) => {
 }
 
 export default TripCard
-
-const Card = styled.div `
-width: 370px;
-height: 300px;
-background-color: rgba(0,0,0,0.9);
-margin: 20px;
-position: relative;
-
-    h2 {
-        position: absolute;
-        padding: 5px;
-        top: 0px;
-        left: 0px;
-        background-color: rgba(0,0,0,0.5);
-    }
-
-img {
-    width: 100%;
-    height: 60%;
-    cursor: pointer;
-    :hover {
-        opacity: 0.8;
-    }
-}
-`
-const Info = styled.div `
-width: 100%;
-height: 30%;
-padding: 5px 10px;
-
-color:#fff;
-
-`
