--- semana10/LabeX/labex/src/hooks/headlesInput.js
@@ -1,8 +1,0 @@
-const { useEffect, useState } = require("react")
-
-import React, {useEffect, useState} from 'react'
-
-const headlesInput = () =>{
-    
-}
-
