+++ semana10/LabeX/labex/src/hooks/useForm.js
@@ -5,15 +5,12 @@ export const useForm = (initialState) => {
 
     const onChange = (name, value) => {
         const newForm = {...form, [name]: value}
-    
         setForm(newForm)
     }
 
     const resetState = () => {
         setForm(initialState)
     }
-
-
 
     return { form, onChange, resetState}
 }
