+++ semana10/LabeX/labex/src/screen/DashBordPage/DashBordPage.jsx
@@ -1,8 +1,7 @@
-import React, {useEffect, useState}from 'react'
+import React, {useEffect}from 'react'
 import {useHistory} from 'react-router-dom'
-import styled from 'styled-components'
-
-import {goToLogin, goToDashBord, goToDetail} from '../../Router/goToPages'
+import {DashBordContainer} from './styled'
+import {goToLogin, goToDetail} from '../../Router/goToPages'
 import TripPage from '../TripPage/TripPage'
 
 const DashBordPage = () => {
@@ -20,10 +19,9 @@ const DashBordPage = () => {
 
 
     const details = (event) =>{
-            goToDetail(history, event.target.id)
+        goToDetail(history, event.target.id)
     }
 
-    
     return (
         <DashBordContainer>
             <h2>Viagens</h2>
@@ -34,12 +32,6 @@ const DashBordPage = () => {
 
 export default DashBordPage
 
-const DashBordContainer = styled.div`
-background-color: #fff;
-width: 100%;
-height: 100%;
-box-sizing: border-box;
-`
