+++ semana10/LabeX/labex/src/screen/DashBordPage/styled.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+export const DashBordContainer = styled.div`
+background-color: #fff;
+width: 100%;
+height: 100%;
+box-sizing: border-box;
+`
