+++ semana10/LabeX/labex/src/screen/TripDetails/TripDetails.jsx
@@ -1,32 +1,31 @@
 import React, { useEffect, useState } from 'react'
 import { useHistory, useParams } from 'react-router-dom'
-import {getTripDetail} from '../../services/requestApi'
-import styled from 'styled-components'
-import { goToLogin } from '../../Router/goToPages'
-
+import {getTripDetail, decideCandidate} from '../../services/requestApi'
+import { goToLogin} from '../../Router/goToPages'
+import {Details,DetailsContent,CandidateInfo,ContainerItem,CandidatesList,BtnTrue,BtnFalse} from './styled'
 
 const TripDetails = (props) => {
     const [ trip, setTrip ] = useState({})
-    const [ candidates, setCandidates ] = useState([])
+    const [ candidateList, setCandidates ] = useState([])
     const [ contentInfo, setContentInfo] = useState(false)
     const [ candidate, setCandidate] = useState({})
+    
     const parm = useParams()
     const history = useHistory()
 
 
     useEffect(() => {
         const token = window.localStorage.getItem("token");
-    
         if (token) {
             getTripDetail(setTrip, setCandidates, parm.id)
         } else {
             goToLogin(history)
         }
 
-    }, [history]);
+    }, [history, candidateList]);
 
     const info = (index) => {
-        setCandidate(candidates[index])
+        setCandidate(candidateList[index])
         setContentInfo(!contentInfo)
     }
 
@@ -41,24 +40,32 @@ const TripDetails = (props) => {
                 <h3>{trip.planet} dia: {trip.date}</h3>
                 <p>{trip.description}</p>
                 <small>Duração da viagem: {trip.durationInDays} dias</small>
-                {/* <button onClick={tripDelete}></button> */}
+
+                <br/>
             </Details>
 
             <CandidatesList>
                 <h1>Candidatos</h1>
-                {candidates.map((candidate, index)=>{
-                    return<Item onClick={()=>info(index)}>{candidate.name}</Item>
-                })}
+                {candidateList.map((candidate, index)=>{
+                    return(
+                        <ContainerItem>
+                            <p onClick={()=>info(index)}>{candidate.name}</p>
+                            <BtnTrue onClick={()=>decideCandidate(true, parm.id, candidate.id)}>Aprovar</BtnTrue>
+                            <BtnFalse onClick={()=>decideCandidate(false, parm.id, candidate.id)}>Recusar</BtnFalse>
+                        </ContainerItem>
+                        )
+                    })}
+            
             </CandidatesList>
 
             {contentInfo && 
                 <CandidateInfo>
                     <h1>Informações</h1>
                     <p>{candidate.name}</p>
-                    <p>{candidate.age}</p>
+                    <p>{candidate.age} anos</p>
                     <p>{candidate.country}</p>
                     <p>{candidate.profession}</p>
-                    <p>{candidate.applicationText}</p>
+                    <p>Informações: "{candidate.applicationText}"</p>
                 </CandidateInfo>
             }
 
@@ -69,40 +76,3 @@ const TripDetails = (props) => {
 
 export default TripDetails
 
-const DetailsContent = styled.div `
-background-color: #ffa;
-display: flex;
-align-items: center;
-margin: 10px;
-`
-const Details = styled.div `
-background-color: #cff;
-width: 30%;
-height:450px;
-padding: 10px;
-box-sizing: border-box;
-`
-const CandidatesList = styled.div `
-background-color: #caf;
-width: 30%;
-height:450px;
-padding: 10px;
-box-sizing: border-box;
-`
-const Item = styled.p `
-background-color: #fff;
-width: 90%;
-padding: 10px;
-box-sizing: border-box;
-cursor: pointer;
-:hover{
-    background-color: #eee;
-}
-`
-const CandidateInfo = styled.div `
-background-color: #1af;
-width: 30%;
-height:450px;
-padding: 10px;
-box-sizing: border-box;
-`
