+++ semana10/LabeX/labex/src/screen/TripPage/styled.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components'
+
+export const TripContainer = styled.div`
+background-color: #fff;
+display: flex;
+flex-direction: column;
+justify-content: center;
+align-items: center;
+`
+export const TripGrid = styled.div`
+background-color: #a8dadc;
+display:flex;
+flex-wrap: wrap;
+justify-content: center;
+`
